Added ciphersuite.py To Query ciphersuite.info.
Taking the output from testssl.sh, ciphersuite.py will query
https://ciphersuite.info for each cipher and return the security rating.
It can also open the weak/insecure ciphers in browser so the reason(s)
behind the rating can be viewed. Ideally, this wouldn't be needed and
this could highlight the weaknesses, e.g. using CBC, but that might be
some future work.