Added reinstall_altmail.sh.
This script is super sketch but it will redownload a known-working copy
of the AltStore Mail.app plugin and reinstall it. Currently, the
Mail.app plugin that's bundled with AltStore doesn't work and gets
unloaded when opening Mail.app. This script should fix that.

The version of the Mail.app plugin is pulled down directly from the
AltStore repo.